requirements.txt
pydantic
langchain-pinecone
langchain
langchain-community
ipykernel
langchain-core
langchain-text-splitters
langgraph
sentence-transformers
python-dotenv
gradio
langchain-huggingface
pinecone-client
requests
pillow
pypdf
pinecone-client